[X] - minor casting changes

+++ X_Proj/HelloX/WinMain.cpp
@@ -2,8 +2,8 @@
 
 bool HelloX::GameLoop(float deltaTime)
 {
-	float mouseX = (float)X::GetMouseScreenX();
-	float mouseY = (float)X::GetMouseScreenY();
+	float mouseX = static_cast<float>(X::GetMouseScreenX());
+	float mouseY = static_cast<float>(X::GetMouseScreenY());
 	X::Math::Vector2 destination(mouseX, mouseY);
 
 	if (X::Math::Distance(scvPosition, destination) > 5.0f)
@@ -13,8 +13,8 @@ bool HelloX::GameLoop(float deltaTime)
 		scvAngle = atan2(-direction.x, direction.y) + X::Math::kPi;
 	}
 
-	int frame = (int)(scvAngle / X::Math::kTwoPi * 16.0f) % 16;
-	X::DrawSprite(scvTextureIds[frame], scvPosition - X::Math::Vector2(32.0f, 32.0));
+	int frame = static_cast<int>(scvAngle / X::Math::kTwoPi * 16.0f) % 16;
+	X::DrawSprite(static_cast<int>(scvTextureIds[frame]), scvPosition - X::Math::Vector2(32.0f, 32.0));
 
 	return X::IsKeyPressed(X::Keys::ESCAPE);
 }
